Add successor proposal button and form.
When a proposal is decided, a new button "Create successor proposal"
appears.
The button leads to a prefilled proposal add form: the defaults of the
form base on the decided proposal.
The original attachments as well as the excerpt are proposed to be
added as attachment.
The original proposal document can be selected as template for the new
proposal.