Add upgrade step to fix proposal history
The history of an object is stored in IAnnotations.
The type of the entry (ex. submitted) is used to filter the entries in
the proposal history. (See https://github.com/4teamwork/opengever.core/blob/master/opengever/meeting/proposalhistory.py#L42)
The key under which the type is stored is called 'history_type'.

Previously the key used was 'name', but changed in 21f89be.
There was no upgrade step (data migration) to change these keys for
existing data.