Create missing SQL sequences.
When creating a new table with a sequence in a schema migration, the
sequence is not created automatically.
We haven't done that up to now.

This change introduces a helper "ensure_sequence_exists", which safely
creates a sequences when it is missing.

An upgrade step makes sure all currently known sequences exist.

The last two upgrade steps were updated to ensure the their sequences
exist in order to improve the copy/paste experience for future migrations.