SPV word: Protect proposal transitions to require modify permission.
Executing a proposal transition requires the "Modify protal content"
permission, but the transitions were listed for users without this
permissions.

With this change we no longer show transitions to users without the
"Modify portal content" permission.

This cannot be soled within the workflow implementation because the
workflow object has no reference to the Plone object (which knows
about security).
It has a reference to the so-called "model"-object, but this object
does not know about security and the Plone object may or may not be
resolvable (e.g. multi client setup where proposal and submitted
proposal are stored in different admin units).