Remove link to dossier if permission is missing
If a user does not have the permission to view the dossier a link to the
dossier is not reasonable.
Therefore when the dossiers are listed, the view permissions are checked
and if the user is missing the permission to view the dossier only the
title of the dossier is rendered.