Schema dumps: Refactor JSON schema generation:
Clean up the generation of JSON schemas (for bundles as
well as API docs) by introducing a new JSONSchema helper
class that handles building and manipulating JSON schemas.

This refactoring does not cause any functional changes to
the dumped schemas.

There are some minor changes to the order of some dictionary
keys, which will be updated in the next commit.