Update schemadumps after stabilizing order (previous commit):
This update only contains changes in the ordering of dictionary
keys. Committed separately from the change so we can have proof
that the tests that check that schema dumps are up to date
(which are order-insensitive) still pass, and the previous
commit didn't change *anything else*.

The previous commit homogenizes the key order for top-level
schema keys as well as property keys. Before, key order in
regard to the 'required' key was different for the main
content type schemas and sub-schemas included in the bundle
schemas.