Do not render document link without View permission.
When a user does not have the permission to view a document, the
DocumentLinkWidget should not render a link (which will lead to
unsufficient privileges), but render a text instead.

This does not affect regular listings, where the users doesnt see such
documents anyway, but explicit relations where the user must know that
a document exists, but may not have access to the document.