Add dossiertemplates configuration flag `respect_dossier_depth`.
Currently dossiertemplates are not respecting the maximum dossier
depth configuration option. The current behavior is the desired one,
for most of the customers, but the behavior should be
configurable. Therefore we add an additional registry option
`respect_dossier_depth` to the dossiertemplate settings. Which is
disabled by default.

See https://basecamp.com/2768704/projects/14549453/todos/326699123 for
more information.