Remove bcc from attributes for resolved dossiers
When a dossier is resolved, the user does not have the permission to
modify its content.
The send as attachment action adds a bcc email address, such that the
document is added to the dossier again.
This should not happen when the dossier is resolved.
Therefore the bcc is removed when the user has no permission to modify
its content.