Fix batch-handling in showroom overlay for the search-view.
The showroom search-view implementation for fetching new objects will use the default
next/previous button urls of the searchform to request new objects based on the current search
params. If we load only the first next or previous batch-slice, this will work. But if we
reach the end/beginning of the second batch-slice, it will reuse the same fetch-link as for
the first request because the search-view itself will not update its buttons after fetching
new objects through the showroom.

To solve this issue, we implement a new class for handling the current request-urls for the
tail/head function of the showroom.