Fix handling of disabled or not assigned users in the sources.
Currently disabled or no longer assigned users has been
invalid. Therefore it was not longer possible to edit a dossier with a
not assigned users defined as responsible, because it raised an
NoResultFound exception.

With this PR disabled or not assigned users are still valid but not
selectable via the search method.