Fix zip export when agendaitemlisttemplate missing
The creation of the zip file fails when there is no agenda item list
template.
The export itself should work when there is no agenda item list
template, the agenda item list should not be included in the zip file in
that case.
Since getting the agenda item list template is obfuscated, a try-except
is used to prevent the server error.