Allow adding excerpt headers and suffixes
A sablon template can be defined on committee or committee container
which will be added (if available) to the created excerpts.
If the committee has no templates available, the template of the
committee container is used.

Since the new excerpts do not only contain the content of the agenda
item document, the tests which previously compared the content of the
generated excerpts with the agenda item documents need to be changed.

The import of the MergeDocxExcerptCommand inside the method of the
AgendaItem is required to get rid of circular imports.