Fix textfilter in sqlsource table listings for oracle backends.
Oracle raises an "ORA-00906: missing left parenthesis" erorr when
casting a integer column to a string and querying it with ilike.

Therefore I add a special handling for oracle backends, which solves
the issue https://sentry.4teamwork.ch/sentry/onegov-gever/issues/3749/
when filtering in task listings like the MyTasks tab.