Fix syncing the submitted proposal title to sql.
The title and other proposal-model attributes should only be synced after
they have been set from the attribute data provided as JSON in the request.

We have not noticed this is a problem so far because the test tests with
attachments which cause an object modified event to be fired later. Same for
the word implementation since creating the word document fires a modified
event, too.