Fix renaming of agenda items in the meeting view
When editing agenda items in the meeting view, the title of the agenda
item appeared doubled in the input field.
Also sending the form only changed the title in part of the view.

The input field needs to be prefilled with the title of the agenda item.
Since there are several types of agenda items (agenda items from
proposals, paragraphs, etc.) the location of the title may vary.
The selector '.proposal_title>a, .proposal_title' matches all title
containers, but returns several elements in the case of agenda items
from proposal items.
In this case, only the first of both titles is used.

The code responsible for rendering the title of the agenda item used the
values from the SQL-Index.
Since the SQL-Index was not updated after changing the title of the
agenda item, the view did not reflect the changes.
Since the index should be used as index only, the title displayed item
is determined from plone content.