Fix dossier responsible widget, in the DossierAddFormView step.
When accepting a multi adminunit task and selecting the store and
editing in a newdossier step. The dossier responsible field does not
work. The AJAX call tot the search endpoint ist not working (404).

The reason is that the typename, which is used by the
WizzardWrappedAddForm, is not set in the widget_traversal request,
therefore I add an fallback to the default dossier type.