Use widget to inject initial data.
Set the initial value on the widget with a data convertor
instead of just injecting the value into the request. This
is a better way of loading data especially since the data is
set from a data converter which can also deal with
vocabularies that have values that differ from their tokens.