Implement cross tab logout mechanism.
When a logout is triggeret in one gever tab, all other tabs are reloaded
automatically.
This is done through a SharedWorker. All tabs are connected with this
SharedWorker. If one tab sends a logout message to the worker the worker
broadcasts a logout message to all open gever tabs.

According to https://caniuse.com/#feat=sharedworkers this is just
working in Chrome and Firefox.