Fixture: prevent overlapping freezed times in fixture segments.
The fixture is segmented into hours where things happen.
Each call to ftw.builder's create function will move the frozen time
forward.
When a segment calls the builder more than 29 times, the frozen time
is in the next hour, which may already be controlled by the next
segment.
So time will go backwards and we have a mess in the intids, which are
patched to be based on the current time.

An assertion on segment tear down will verify that the segment didn't
took too long.

Some segments are split up into more segments in order to fix such
issues and give us more room in the future.