Improve proposals/decisions in protocol generation
The protocol is generated using a protocol header and a footer and
adding all proposal documens between the header and the footer.

The protocol needs to contain the agenda item number before each
proposal.
It also needs to include the title of the agenda item.
Further styling desires (like a page break after each agenda item) is
also accomplishable by adding these templates.