Introduce system-wide tracking of schema migrations.
Schema migrations must only be applied once per system (=SQL-db).
In an environment with multiple admin units (=Plone sites) it is
therefore necessary that only the first admin unit running upgrades
runs schema migrations and the other admin units are not running the
same schem migrations again.

This feature was implemented by storing the target version of the last
executed schema migration per profile in the SQL
database (opengever_upgrade_version table).
The problem with this approach is that it does not support orphan
schema migration because it does not track the installation for each
schema migration but uses the latest version for comparison.
(An orphan uprade step is one that is not yet installed but has a
lower target version than the currently installed version).

The approach for fixing that is to track all schema migration
separately, requiring a change in the primary key of the table.
The tracking is moved to a GEVER-specific IUpgradeStepRecorder
customization which also stores tracking information the SQL database
for schema migrations, allowing us to actually mark migrations as
executed on other admin units, rather than marking them as proposed
and skipping the execution.

In order for making these changes easy, the knowledge of the
operations class had to be changed so that it does not depend on
upgrade steps.