Implement bumblebee's auto refresh functionality
As bumblebee has the capability to send a notification to a client
using WebSockets when a document has finished, we can now profit from
that, embedding the client library `bumblebee-cable.js`.

This improves the user experience a lot, without any performance costs
on bumblebee. So if the user accesses a site, he'll first get the
thumbnail and then the preview container (either image or iframe) will
automatically be refresh upon completion by bumblebee.