Fix testing cache by dropping opengever_upgrade_version table.
When running functional tests, the opengever_upgrade_version may be
left over because it is not registered in the metadata because of it's
special nature (it is created on the fly).

This leads to problems when running integration tests from cache after
running functional tests: in this case we have the table in our cache
and cannot recreate it because it already exists.

The easiest solution is to just drop the table before loading from
cache.