Add files from solr-conf to data_files: This is needed
so that they get placed into the .egg during installation.
Including them in MANIFEST only ensures these files are
included in the source tarball - but for files *outside*
of Python packages, these need to be explicitely listed
in data_files in order for them to be copied *into the
egg* during buildout time.