Make reference_number field read-only
This prevents setting the field's value using the REST API.

Still keep the field in the edit form, though it's questionable why.
Also questionable, if the reference number should be a schema field at all.

Do not set default values for read-only fields, as this fails.
Also fix some tests that read/write the sequence number in a wrong way.