Temporarily disable upgrade step to fix broken references until we're sure it's safe:
Disable this upgrade step until we figure out what exactly is going
on in issue #4074 - right now it seems that pickling a persistent
object (just to check whether it's in an ok state) might be a bad
idea and not be side-effect free.

Until we're sure this isn't harmful, we disable this upgrade step.