Make sure bumblebee checksum gets calculated for docs created via REST API:
The event handler in opengever.bumblebee that does so on ObjectModified events
relies on doc.digitally_available being up-to-date. This was not the case
because both event handlers are registered for the same event type, and
we can't control the order in which these event handlers are executed.

We therefore must make sure to update doc.digitally_available before
checking against it.