Fix API get request on private dossiers.
Currently API requests on a private dossiers raises an `TypeError: 'str' object is not callable`. Because of an incorrect handling in the `plone.app.contentlisting` the adapted IContentListing-object looses the acquisition context which leads then to problems when getting the owner on in the title accessor of a PrivateFolder.

I fixed the issue by overriding the `RealContentListingObject` adapter. I'll add also an PR to the `plone.app.contentlisting` package, but that needs some more time and a general test-example.