Introduce ObjectTouched event:
This event will cause a "touch" to be logged for the object
in question. These "touches" are stored in a rotating per-user
log, and will be used to display a user's recently touched
objects.

The semantics of what exactly constitutes a "touch" is
controlled by when this event is fired. For the current
implementation, we're considering significant writes,
e.g. ObjectAdded, ObjectModified, checked in/out, ...