Fix resolving tasks when original was delegated
When working with multi admin setups, tasks may span over several admin
units.

We assume to have two admin units: FD and RK

Anton creates a task in FD and gives it to Berta in RK
At the same time, Anton delegates the task to Carlos and David in FD
When Berta tries to resolve the task, the subtasks delegated to Carlos
and David should be checked for completeness as well.

Therefore the check on Berta's side should get all the subtasks from
Berta's and Anton's task.