Allow to select custom template for ad hoc agenda items
This was solved using the vocabulary, since the allowed templates on the
committee may be empty and therefore None.
The vocabulary always returns the right templates (all if None is
selected).

The vocabulary is imported inside the else, because importing it at the
top of the file leads to a circular import error.

The template_id is passed as optional argument.
If no template id is given, the default template for ad-hoc agenda
items is chosen.