Fix accessing mail file in zip export
Use get_file instead of file to access the file objects in mails as
well.
This is required when proposals have mails as attachments.

Also:

This test will not work, because of versioning and SQL.
When an attachment is added to the proposal, the meeting application
should display a message, reminding the user of the updated proposal.

Therefore the documents available in the proposal where the ones before
adding the email to the proposal.

```python
@browsing
def test_meeting_can_be_exported_to_zip_when_mail_within_submitted_proposal(self, browser):
    self.login(self.committee_responsible, browser)

    create(Builder('mail')
           .with_message(u'Fun m\xfcil')
           .within(self.submitted_proposal))
    self.schedule_proposal(self.meeting, self.submitted_proposal)

    browser.open(self.meeting, view='export-meeting-zip')
    statusmessages.assert_no_error_messages()
    self.assertEquals('application/zip', browser.contenttype)
```