Fix updating mail title via API.
Currently mail titles gets resetted to subject, when trying to update
a mail via REST-API PATCH requests. It now only
snycs the title from subject if a message object is part of the
parameters.

This fix makes it also possible to create a mail with a different
title than the subject.