Put comments as well in the title attribute
Comments are truncated when they are too long.
A simple way to solve this without breaking the UI is to put the comment
as well in the title attribute of the DOM element containing the
truncated text.

This is already solved in the Journal tab.
The tooltip helper just strips all html tags from the text, probably to
prevent XSS.