Fix a bug with modern emojis in task responses.
Fix a bug with modern emojis while rendering task responses. The source of the
problem was not calling `getData` on a transform  which will return a stream
instead of the result.

Testing seems a bit difficult at the moment as our tooling does not seem to
support these characters either 😂.