Tests: Untangle admin units and org units in FixtureBuilder for functional test cases:
Both the admin unit as well as the org unit in the Fixture builder
were created with the ID 'client1'. In addition, there was some code
that attempted to link the org unit to the admin unit, but all it really
did was *set the ID of the admin unit to the ID of the org unit*.

Because both they're IDs were 'client1' by default, that didn't cause
any harm in cases where the admin unit ID wasn't customized.

This change untangles their IDs, setting them to 'org-unit-1' and
'admin-unit-1' respectively.