Create initial version with the original document creator as the principal:
With the deferred creation of the initial version, the initial version will
only be created once a second version is actually created. That may happen
in the security context of a different user that initially created the
document.

We therefore create the initial version with a security context that assumes
the user ID of the document's initial creator to make sure the 'principal' in
the initial version's authorship metadata is correct.