Convert pyxb dates to real python datetime dates.
With the plone.restapi version and the corresonding default values
patch, pyxb date's are now passed in correctly but raises an
ValidationError because of the datetime isinstance check in
zope.schema.field(279).

Maybe this could be manged by the pyxb binding in ech0147 bindings,
but I need to discuss this with buchi.