Make sure widgets properly return persisted missing values
for optional fields with default values:

In z3c.form.widget.Widget.update there is a misguided check that falls
back to using a default value if a fields actual value is equal to
missing value.

This leads to a bug where for fields that are

1) optional (not required)
2) have a default of some sort (field.default or field.defaultFactory)
3) have a legitimate missing value persisted
4) have a MV that is different from the default

an incorrect value is returned by the widget (the default value instead
of the legitimate missing value).

This will result in display views that render an object and use the
widget's display mode for that to render an incorrect field value
(even though underneath, on the persistent object, the MV is
properly persisted).

In addition, in an edit form the widget will incorrectly cause
the default to be filled in instead of leaving the field at its
legitimate MV. This will then result in the persisted MV being
overriden by the default.

This change monkey patches z3c.form.widget.Widget.update in a way
where this strategy is only applied for REQUIRED fields: Only
if the field is required AND the persisted value is equal to MV
will it return the default - otherwise, the persisted value is used.