Overwrite showEditableBorder from ploneview,
to respect also worfklow transitions.

This solves an issue, where the contentmenu was wrongly not visible,
in some cases where the user is able to trigger a transition, but does
not have Modify oder Add portal content permission. For example an
administrator is allowed to reactivate a inactive dossier, but did not
see the menü.