Improve testing of dossier resolve conditions.
Testing of the dossier resolve conditions was weak as
we did not actually test that the dossier did not get resolved
when conditions were not met. One of the tests was broken, so that
it was not tested that a dossier containing subdossiers with invalid
end dates could not be resolved.