Make upgrade step 20180619143343 deferrable:
The use of CMFEditions' CopyModifyMergeRepositoryTool.getHistory()
causes massive amounts of savepoints (one per version history retrieve).

When this is done for lots of documents, this leads to thousands of
savepoints being created and rolled back to in the same transaction.

This has the side effect of creating just as many nested SQLAlchemy
transactions.

When the SQLAlchemy connection then is used in a subsequent
upgrade step, the attempt to fetch the outermost connection in
sqlalchemy.orm.session.connection_for_bind() then fails,
because it's recursive and exceeds Python's maximum recursion
depth (1000 by default) because of the many nested transactions.

We therefore mark this upgrade step as deferrable, so that it can
be run seperately from other upgrades, hopefully sidestepping this
issue.