Move the favorite action just after the documentFirstHeading.
The favorite-action is rendered by the below-content-title
viewlet, but we actually want it on the same line as the title.
This viewlet is rendered as a block in the default template, below
the title. We use diazo here to move it outside of that viewlet
and just after the document title.