Generate zip on last demand callback.
Previously the zip would be generated on each download. We change that behavior
and generate the zip as soon as the last demand callback arrives. This trades
off blocking a bumblebee request vs always regenrating the zip when the button
is clicked. As the zip may live for some hours/days and be downloaded multiple
times we chose to only generate the zip once.