Don't store UUID objects and convert back and forth between them
and their string representation - directly store UUID strings
instead.

This also removes the validation for conformance to a particular
UUID4 format - it brings us no value and is not an expected error case.
All we really care about is whether there's a job with that particular ID
or not - the exact contents and format of the ID are irrelevant.