Add upgrade step to reindex allowedRolesAndUsers in Solr:
This is needed to ensure that the allowedRolesAndUsers index in
Solr is up-to-date after bumping ftw.solr to a version that also
patches reindexObjectSecurity().

In addition, we optimize an upgrade step from the past that did
exactly this, so we don't reindex allowedRolesAndUsers in Solr
twice (at least for deployments where 2018.4 was never installed).