Fix local role assignment for meeting content.
Use correct roles for meeting content. This fixes having to assign roles
manually for development. It also fixes incorrect assignments for dev content
on committee container level that could have weird side-effects.

Furhermore explicitely list the contes for which to assign default roles to
prevent this in the future and to make more obvious what happens.