Fix rendering issue in search view.
Fix it for solr and for the skins folder search results. The description was
inserted as structure which would not render correctly due to missing escaping
for catalog results.

SOLR search is configured by the view to return markup with highlihting using
<em> tags. SOLR is now configured to already escape its results before inserting
those <em> tags so that the returned description can be safely inserted in the
template using a tal:structure.

For catalog results the view now escapes the description before it is being
rendered in the template.