Fix issue when returning an excerpt to an already decided proposal.
This will not happen in the long run, but agendaitems that were
decided before we introduced that proposals get decided only when
the excerpt is returned can be already decided even if the proposal
has not been returned. This lead to an error when returning the
excerpt and trying to decided an already decided proposal.