Add Open XML Visio to mimetype registry.
This upgrade step is in opengever.core and not in opengever.policy.base to
prevent ordering issues as this upgrade *must* be installed before the next
one which reindexes content types to have correct css classes.

We will drop the separate mimetypes profile in the forseeable future and it is
guaranteed that it is always installed together with opengever.core even though
there is no dependecy in `metadata.xml`.