Make ReindexTaskPrincipalsInGlobalindex upgradestep more robust.
Skip tasks where no representing sql task is storred in the
globalindex. This situation should never happen and is something which
should be visible in the validation scripts, but should not let the
upgrade fail.

So we skip and log a warning instead if a corresponding sql task is missing.