Return review_state ID in API summaries
And introduce a new review_state_label attribute instead. The
`review_state` of summaries should work in the same manner as it does
for a object representation of a GET requests. For applications which
need the translated label of the review state we introduced an
additional attribute `review_state_label`.