Add role assignments to referenced proposals.
Add role assignments to a proposal whose document is referenced by a task. But
only do so if the document itself already has local role assignments from the
task.

Instead of iterating over all tasks or documents we start at the proposal side
assuming that there are fewer proposals than tasks or documents. Also that way
the migration will only run for deployments that have meeting content, so there
is no need to check against the feature flag.