Fix file indexes upgradestep for deployments using solr.
The change from PR #5024 only added catalog indexes, but did not add
corresponding indexes to the solr schema. So the upgradestep was
useless for deployments with an enabled solr search or even worse it
almost removed all solr data because of a curios behave of solr or
ftw.solr. I'll investigate this behavior in a separate issue.

So I added the new indexes to the schema and fix the current upgradestep.