Fix create zip for concurrent callback requests.
When callback requests arrive concurrently the they may not correctly see the
amount of finished documents, so zip creation was not triggered all the time.
Thus we move triggering creating the zip file to the polling view. The polling
view should only be called once as the polling interval is quite long. It can
also deal with two concurrent zip creations though and also does not create
the zip again when a zip is already present.