Set changed date on ObjectAdded instead of ObjectCreatedEvent.
This fixes an issue when copy & pasting a dossier containing
documents. The reason is, that object which the CreatedEvent or
CopiedEvent (inherits from the CreatedEvent) aren't acquisition
wrapped. This can lead to several issues, as example the copy & paste
bug.

IMHO the ObjectAdded which is fired when the object is added to the
context, is the right thing to initialize the changed date.