Remove plonetheme.teamraum upgradesteps.
Those upgrades are outdated and no longer used, but can lead to
problems, because the plonetheme.teamraum upgrades are run after the
opengever.core ones. For example the upgradestep which removes the
`++theme++plonetheme.teamraum/js/dropdown.js` (`RegisterMissingDropdownJS`)
is run before the older plonetheme.teamraum upgradestep which adds the
`dropdown.js`. So after an upgrade a wrong and not working dropdown.js
is registered.