Fix loading of document preview with tooltip
As documented in https://extranet.4teamwork.ch/support/stadt-nidau/tracker-gever/95 it may be possible that the documents overview page reloads when loading a tooltip of a document and the document detail view simultaneously.
So when user clicks on the document link in the overview while the browser awaits the response from the tooltip, the tooltip request is aborted. The javascript code checks the response of the tooltip request, which has been aborted, for a `X-Tooltip-Response` header. But because the request has been canceled the header is not present. As a result the javascript code triggers a reload of the documents overview page because it is treated as an error.
Using a flag setting to true when the user clicks on the documents links solved the problem. The flag is checked before reloading the page.
Also adjust the error message when the `X-Tooltip-Response` header is missing.