WebActions storage: Add support for unique_name field:
The unique_name field is an optional, client-controlled value
that can be used by clients to more easily avoid creating the
same webaction twice, without having to track the server-generated
action_id.

If a unique_name is specified, the server will reject creating the
action if one with the same unique_name already exists.