Don't truncate long groupid tokens.
The normalizer normalizes to a length of 50 by default. As this is shorter
than the maximum length of the groupid column and some customers have a naming
schema where the the first 50 chars of a group id might be identical we must
allow for longer tokens. Same maximum length as the groupid column should be
safe.