Solr TabbedView filters: Also include non-wildcarded terms in query:
Wildcard queries in Solr only get analyzed with a subset of the query
analyzers, not all of them (only multi term aware ones).

Specifically, this leads to our WhitespaceTokenizerFactory not being
run against the query, and the query therefore doesn't get tokenized.

Also including the non-wildcarded terms in the query pattern allows
us get better matches for these cases where the user enters a search
term that needs to be tokenized.