Fix preview tooltip for different input types.
Closes https://github.com/4teamwork/opengever.core/issues/4809

Use a flag which will be set to true when the user touches the screen. So when the mouseover event is triggered after the touchevent, the tooltips are not initialized. On the other hand when user does not touch the screen, the touchevent is not fired so the tooltip renders as expected.