Reenable testserver selftest on jenkins by fixing port choosing.
The port choosing method is switched to using a 0-port so that the OS
chooses the port in the testserver.

This is done in order to avoid problems where the jenkins master
chooses a port, but the test is run at the slave, on which the port is
already used.

The robot server is extended so that it prints out the URLs including
the chosen ports for the ZServer and the XMLRPC-server.
The selftest script now needs to watch the stdout of the testserver
process and wait until the URLs appear.