Avoid infinite loops when looking for parent dossiers:
When using while loops to walk up the parents in the acquisition chain
to find a particular object (or stop at the site root), care must be
taken to consider the case of improperly AQ wrapped objects:

When these functions get passed objects that aren't AQ wrapped (or
don't have a full acquisition chain that reaches up to the Plone
site), aq_parent() will return None at some point.

If this isn't accounted for and dealt with by terminating the loop,
this kid of a while loop can end in an infinite loop.