Suppress deletion events when filtering objects from copied subtrees:
When "filtering" objects (currently tasks and proposals) from copied
subtrees, we do so by deleting them from the temporary ZEXP that's
about to be attached to the destination container.

We need to suppress deletion events in order to avoid attempts to
uncatalog objects that 1) haven't even been cataloged yet and
2) dont' have a proper AQ chain because they're part of a subtree
that hasn't been attached to a container yet.