Lock dossier subtree during resolve transition:
Implements a locking mechanism to prevent concurrent dossier resolution.

This mechanism is intended to prevent users from simultaneously triggering
the 'resolve' transition for a dossier (which would result in degraded
performance because retries of conflicting transactions).

We do this by issuing a persistent lock that gets committed in its own
transaction as the very first thing in the resolution process. Further
attempts at resolving a dossier are then rejected as long as such a lock
exists (and hasn't expired). Once the dossier is resolved successfully,
the lock is removed and the removal will be committed again.

In case of an exception, we catch it, abort the transaction, remove the
lock, commit the removal, and re-reaise the exception to be handled by
the usual error handling in ZPublisher.