Fix performance issue with search root exclusion in tabbed view listings:
The original implementation in ftw.table creates one path query for each
immediate child. Turns out that performs much worse than expected for
containers with lots of immediate children.

Because in GEVER we have a more recent version of Products.ZCatalog, and
also monkeypatch the UIDIndex's query options to add support for the 'not'
operator, we're able to use a more efficient strategy here, and simply
exclude the current context using a {'not': uuid} filter.