Fix solr error during copy/paste of word document.
Docproperties get updated several times during a copy/paste
operation of a word document, overwriting the blob file each
time. Each time the blob file is changed, a reindex operation
is queued in solr, which is processed after the transaction ends.
When that happens, solr tries to process some blobs that do not
exist anymore and fails. We now update the docproperties only
once during a copy/paste operation.