Add test suite for dossier WF transitions via REST API:
This test suite exercises the possible dossier workflow
transitions via the REST API.

It does so on a very high level, for the happy path, and doesn't assert
on all the aspects of the resulting state, post-transition jobs etc.

More detailed tests for the specific transitions can be found in
opengever.dossier.tests.