Make existing IndexForBlockedLocalRoles upgradestep deferrable.
Originally this upgradestep was not implemented as an deferrable
Ugpradestep, but because the blocked_local_roles is only used for an
admin view, we did this now to safe time on large deployments. To make
sure the index exists in any case, I added an additional non-deferrable upgradestep
which only adds the index, if it's not already existing. This
upgradestep is new and will be run also on already updated
deployments, but is skipped because the index already exists.