Nightly jobs: Simplify job interface by eliminating NightlyJob class:
Instead of a wrapper object, jobs are now simply defined as a dictionary
of the arguments that uniquely define a job. The provider which a job
came from is also not stored on the job object any more, but it's the
responsibility of the runner to do bookkeeping for that.