Fix filtering of trashed documents in search results:
The filter that omits trashed documents is only supposed to be
added if no explicity constraint against the 'trashed' index is part
of the query.

This was checked for in the `kw` query dict, but `REQUEST` may also
contain a query. REQUEST may not necessarily be a HTTPRequest, but
can also be a dict. This is part of some older ZCatalog APIs.

We therefore only add the 'trashed' filter if there isn't one
in either `kw` or `REQUEST`.