Docs: Update nightly cron jobs example:
The nightly cron jobs need to run with the zope user's
environment variables. Because these are not normally
sourced in a crontab, we prefix the nightly job command
with `bash -l -c` in order to run them in a login shell,
which will load the environment variables.

This is of particular importantance for the nightly jobs
for at least two reasons:
- When triggering archival PDF generation via bumblebee,
  the http_proxy and no_proxy environment variables must
  be set correctly, otherwise requests may fail (with
  502 notresolvable / BadStatusLine errors).
- We're attempting to log errors from the nightly jobs to
  Sentry. For that to work, the RAVEN_DSN env variable
  needs to be set properly.